+++ diskover_qumulo.py
@@ -136,10 +136,13 @@ def qumulo_api_walk(top, ip, ses):
 
 
 def qumulo_treewalk(path, ip, ses, q_crawl, num_sep, level, batchsize, cliargs, reindex_dict, bar):
+    jobs = []
     batch = []
+    emptydircount = 0
 
     for root, dirs, files in qumulo_api_walk(path, ip, ses):
         if len(dirs) == 0 and len(files) == 0 and not cliargs['indexemptydirs']:
+            emptydircount += 1
             continue
         if root['path'] != '/':
             root_path = root['path'].rstrip(os.path.sep)
@@ -149,7 +152,8 @@ def qumulo_treewalk(path, ip, ses, q_crawl, num_sep, level, batchsize, cliargs, 
             batch.append((root, files))
             batch_len = len(batch)
             if batch_len >= batchsize:
-                q_crawl.enqueue(scrape_tree_meta, args=(batch, cliargs, reindex_dict,))
+                job = q_crawl.enqueue(scrape_tree_meta, args=(batch, cliargs, reindex_dict,))
+                jobs.append(job)
                 del batch[:]
                 if cliargs['adaptivebatch']:
                     batchsize = adaptive_batch(q_crawl, cliargs, batchsize)
@@ -176,7 +180,8 @@ def qumulo_treewalk(path, ip, ses, q_crawl, num_sep, level, batchsize, cliargs, 
                 bar.update(0)
 
     # add any remaining in batch to queue
-    q_crawl.enqueue(scrape_tree_meta, args=(batch, cliargs, reindex_dict,))
+    job = q_crawl.enqueue(scrape_tree_meta, args=(batch, cliargs, reindex_dict,))
+    jobs.append(job)
 
     # wait for queue to be empty and update progress bar
     while True:
@@ -197,6 +202,8 @@ def qumulo_treewalk(path, ip, ses, q_crawl, num_sep, level, batchsize, cliargs, 
         if q_len == 0 and workers_busy == False:
             break
         time.sleep(.5)
+
+    return jobs, emptydircount
 
 
 def qumulo_get_dir_meta(worker_name, path, cliargs, reindex_dict, redis_conn):